+++ packages/devtools/test/support/test_utils.dart
@@ -1,0 +1,17 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+import 'dart:convert';
+
+import 'package:devtools/src/timeline/timeline_protocol.dart';
+
+TimelineEvent testTimelineEvent(Map<String, dynamic> json) =>
+    TimelineEvent(testTraceEventWrapper(json));
+
+TraceEvent testTraceEvent(Map<String, dynamic> json) =>
+    TraceEvent(jsonDecode(jsonEncode(json)));
+
+int _testTimeReceived = 0;
+TraceEventWrapper testTraceEventWrapper(Map<String, dynamic> json) {
+  return TraceEventWrapper(testTraceEvent(json), _testTimeReceived++);
+}