+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 - TBD
-* Make timeline export format compatible with trace viewers such as chrome://tracing.
+* Make timeline snapshot format compatible with trace viewers such as chrome://tracing.
+* Add ability to import timeline snapshots via drag-and-drop.
 * Memory instance viewer handles all InstanceKind lists.
 * CPU profiler bug fixes and improvements.